Auth done but need to add DTOS to main modules

+++ Controllers/PostsController.cs
@@ -1,6 +1,10 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using ForumAPI.Data.Entities;
 using ForumAPI.Data.Repositories;
+using Microsoft.AspNetCore.Authorization;
+using ForumAPI.Auth.Model;
+using System.Security.Claims;
+using Microsoft.IdentityModel.JsonWebTokens;
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace ForumAPI.Controllers
@@ -11,10 +15,13 @@ namespace ForumAPI.Controllers
     {
         private readonly IPostsRepository _postsRepository;
         private readonly IThreadsRepository _threadsRepository;
-        public PostsController(IPostsRepository PostsRepository, IThreadsRepository ThreadsRepository)
+        private readonly IAuthorizationService _authorizationService;
+
+        public PostsController(IPostsRepository PostsRepository, IThreadsRepository ThreadsRepository, IAuthorizationService AuthorizationService)
         {
             _postsRepository = PostsRepository;
             _threadsRepository = ThreadsRepository;
+            _authorizationService = AuthorizationService;
         }
         /*
          *
@@ -28,6 +35,7 @@ namespace ForumAPI.Controllers
 
         // GET: /api/v1/topics/{topicId}/threads/{threadId}/posts
         [HttpGet]
+        [Authorize(Roles = ForumRoles.AnonGuest)]
         public async Task<IEnumerable<Posts>> Get(int topicId, int threadId)
         {
             var posts = await _postsRepository.GetMultipleAsync(topicId, threadId);
@@ -37,6 +45,7 @@ namespace ForumAPI.Controllers
 
         // GET: /api/v1/topics/{topicId}/threads/{threadId}/posts/{id}
         [HttpGet("{id}")]
+        [Authorize(Roles = ForumRoles.AnonGuest)]
         public async Task<ActionResult<Posts>> Get(int topicId,int threadId, int id)
         {
             var posts = await _postsRepository.GetAsync(topicId, threadId,id);
@@ -52,11 +61,13 @@ namespace ForumAPI.Controllers
 
         // POST /api/v1/topics/{topicId}/threads/{threadId}/posts
         [HttpPost]
+        [Authorize(Roles = ForumRoles.AuthForumUser)]
         public async Task<ActionResult<Posts>> Post(int topicId,int threadId, Posts createpost)
         {
             Threads thread = await _threadsRepository.GetAsync(topicId,threadId);
             createpost.Thread = thread;
             createpost.CreationDate = DateTime.Now;
+            createpost.UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub);
             await _postsRepository.InsertAsync(createpost);
             createpost.Thread = null;
             return Created("", createpost);
@@ -64,6 +75,7 @@ namespace ForumAPI.Controllers
 
         // PUT /api/v1/topics/{topicId}/threads/{threadId}/posts/{id}
         [HttpPut("{id}")]
+        [Authorize(Roles = ForumRoles.AuthForumUser)]
         public async Task<ActionResult<Posts>> Put(int id, int topicId, int threadId,Posts updatePost)
         {
             var post = await _postsRepository.GetAsync(topicId,threadId,id);
@@ -71,7 +83,12 @@ namespace ForumAPI.Controllers
             // 404
             if (post == null)
                 return NotFound();
-
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, post, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 403
+                return Forbid();
+            }
             post.Description = updatePost.Description;
             await _postsRepository.UpdateAsync(post);
 
@@ -80,6 +97,7 @@ namespace ForumAPI.Controllers
 
         // DELETE /api/v1/topics/{topicId}/threads/{threadId}/posts/{id}
         [HttpDelete("{id}")]
+        [Authorize(Roles = ForumRoles.AuthForumUser)]
         public async Task<ActionResult<Posts>> Delete(int id, int topicId, int threadId)
         {
             var post = await _postsRepository.GetAsync(topicId,threadId,id);
@@ -87,12 +105,17 @@ namespace ForumAPI.Controllers
             // 404
             if (post == null)
                 return NotFound();
-
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, post, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 403
+                return Forbid();
+            }
             await _postsRepository.DeleteAsync(post);
 
 
-            // 200
-            return Ok();
+            // 204
+            return NoContent();
         }
     }
 }
