+++ Controllers/ThreadsController.cs
@@ -1,6 +1,10 @@
-﻿using ForumAPI.Data.Entities;
+﻿using ForumAPI.Auth.Model;
+using ForumAPI.Data.Entities;
 using ForumAPI.Data.Repositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.JsonWebTokens;
+using System.Security.Claims;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -12,12 +16,13 @@ namespace ForumAPI.Controllers
     {
         private readonly IThreadsRepository _threadsRepository;
         private readonly ITopicsRepository _topicsRepository;
+        private readonly IAuthorizationService _authorizationService;
 
-        public ThreadsController(IThreadsRepository ThreadsRepository, ITopicsRepository TopicsRepository)
+        public ThreadsController(IThreadsRepository ThreadsRepository, ITopicsRepository TopicsRepository, IAuthorizationService AuthorizationService)
         {
             _threadsRepository = ThreadsRepository;
             _topicsRepository = TopicsRepository;
-
+            _authorizationService = AuthorizationService;
         }
         /*
          *
@@ -31,6 +36,7 @@ namespace ForumAPI.Controllers
 
         // GET: api/v1/topics/{topicId}/threads
         [HttpGet]
+        [Authorize(Roles = ForumRoles.AnonGuest)]
         public async Task<IEnumerable<Threads>> Get(int topicId)
         {
             var threads = await _threadsRepository.GetMultipleAsync(topicId);
@@ -39,6 +45,7 @@ namespace ForumAPI.Controllers
 
         // GET api/topics/{topicId}/threads/{id}
         [HttpGet("{id}")]
+        [Authorize(Roles = ForumRoles.AnonGuest)]
         public async Task<ActionResult<Threads>> Get(int topicId,int id)
         {
             var threads = await _threadsRepository.GetAsync(topicId,id);
@@ -53,11 +60,13 @@ namespace ForumAPI.Controllers
 
         // POST api/v1/topics/{topicId}/threads
         [HttpPost]
+        [Authorize(Roles = ForumRoles.AuthForumUser)]
         public async Task<ActionResult<Threads>> Post(int topicId, Threads createThread)
         {
             Topics topic = await _topicsRepository.GetAsync(topicId);
             createThread.Topic = topic;
             createThread.CreationDateTime = DateTime.Now;
+            createThread.UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub);
             await _threadsRepository.InsertAsync(createThread);
             createThread.Topic = null;
             return Created("", createThread);
@@ -66,6 +75,7 @@ namespace ForumAPI.Controllers
 
         // PUT api/v1/topics/{topicId}/threads/{id}
         [HttpPut("{id}")]
+        [Authorize(Roles = ForumRoles.AuthForumUser)]
         public async Task<ActionResult<Threads>> Put(int topicId,int id, Threads updateThread)
         {
             var thread = await _threadsRepository.GetAsync(topicId, id);
@@ -73,7 +83,12 @@ namespace ForumAPI.Controllers
             // 404
             if (thread == null)
                 return NotFound();
-
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, thread, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 403
+                return Forbid();
+            }
             thread.Description = updateThread.Description;
             await _threadsRepository.UpdateAsync(thread);
 
@@ -82,6 +97,7 @@ namespace ForumAPI.Controllers
 
         // DELETE api/v1/topics/{topicId}/threads/{id}
         [HttpDelete("{id}")]
+        [Authorize(Roles = ForumRoles.AuthForumUser)]
         public async Task<ActionResult<Threads>> Delete(int topicId,int id)
         {
             var thread = await _threadsRepository.GetAsync(topicId, id);
@@ -89,12 +105,17 @@ namespace ForumAPI.Controllers
             // 404
             if (thread == null)
                 return NotFound();
-
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, thread, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 403
+                return Forbid();
+            }
             await _threadsRepository.DeleteAsync(thread);
 
 
-            // 200
-            return Ok();
+            // 204
+            return NoContent();
         }
     }
 }
